+++ main.py
@@ -1,1697 +1,14 @@
 #!/usr/bin/env python3
-"""
-Simple Revers-o: Streamlined Visual Investigation Tool
-GroundedSAM + PE-Core-L14-336 + Simple Vector Search
-
-Simplified based on PE research:
-- Single model: PE-Core-L14-336 (optimal balance)
-- Single layer: 24 (PE research optimal for L14)
-- Core pipeline: Image → GroundedSAM → PE embeddings → Vector search
-- Simple UI: Build database, search similar regions
-"""
-
-import os
-import sys
-import time
-import uuid
-import tempfile
-from pathlib import Path
-import numpy as np
-import cv2
-import torch
-import matplotlib.pyplot as plt
-from PIL import Image, ImageDraw
-import gradio as gr
-from qdrant_client import QdrantClient
-from qdrant_client.http import models
-import shutil
-import hashlib
-import urllib.parse
-
-# Add perception_models to path
-sys.path.append('./perception_models')
-import core.vision_encoder.pe as pe
-import core.vision_encoder.transforms as transforms
-
-# GroundedSAM import
-from autodistill_grounded_sam import GroundedSAM
-from autodistill.detection import CaptionOntology
-
-# Video processing imports
-try:
-    from scenedetect import VideoManager, SceneManager
-    from scenedetect.detectors import ContentDetector
-    VIDEO_PROCESSING_AVAILABLE = True
-except ImportError:
-    VIDEO_PROCESSING_AVAILABLE = False
-    print("[WARNING] Scene detection libraries not available. Install with: pip install scenedetect")
-
-# Check for yt-dlp availability
-try:
-    import yt_dlp
-    YT_DLP_AVAILABLE = True
-except ImportError:
-    YT_DLP_AVAILABLE = False
-    print("[WARNING] yt-dlp not available. URL video downloads will not work.")
-
-# =============================================================================
-# VIDEO PROCESSING FUNCTIONS
-# =============================================================================
-
-def is_supported_video_url(url):
-    """
-    Check if the URL is from a supported video platform.
-    
-    Args:
-        url: URL string to check
-        
-    Returns:
-        bool: True if URL is from a supported platform
-    """
-    if not url or not isinstance(url, str):
-        return False
-    
-    try:
-        parsed = urllib.parse.urlparse(url.strip())
-        domain = parsed.netloc.lower()
-        
-        # Remove 'www.' prefix if present
-        if domain.startswith('www.'):
-            domain = domain[4:]
-        
-        supported_domains = {
-            'youtube.com', 'youtu.be', 'youtube-nocookie.com',
-            'twitter.com', 'x.com', 'nitter.net',
-            'facebook.com', 'fb.com', 'm.facebook.com',
-            'instagram.com', 'tiktok.com', 'vimeo.com',
-            'dailymotion.com', 'twitch.tv'
-        }
-        
-        return domain in supported_domains
-    except Exception:
-        return False
-
-def download_video_from_url(url, output_dir, max_quality='720p'):
-    """
-    Download a single video from URL using yt-dlp.
-    
-    Args:
-        url: Video URL to download
-        output_dir: Directory to save the downloaded video
-        max_quality: Maximum video quality (e.g., '720p', '1080p', 'best')
-        
-    Returns:
-        tuple: (success, message, downloaded_file_path)
-    """
-    if not YT_DLP_AVAILABLE:
-        return False, "yt-dlp not available. Please install it: pip install yt-dlp", None
-    
-    if not is_supported_video_url(url):
-        return False, f"Unsupported URL or invalid format: {url}", None
-    
-    try:
-        # Ensure output directory exists
-        os.makedirs(output_dir, exist_ok=True)
-        
-        # Create a safe filename using timestamp and hash
-        url_hash = hashlib.md5(url.encode()).hexdigest()[:8]
-        timestamp = int(time.time())
-        safe_filename = f"video_{timestamp}_{url_hash}.%(ext)s"
-        
-        # Configure yt-dlp options with more predictable filename
-        ydl_opts = {
-            'outtmpl': os.path.join(output_dir, safe_filename),
-            'format': f'best[height<={max_quality[:-1]}]/best',  # Simplified format to avoid merging issues
-            'merge_output_format': 'mp4',
-            'writeinfojson': False,
-            'writethumbnail': False,
-            'quiet': True,
-            'no_warnings': True,
-            'restrictfilenames': True,  # Use only ASCII characters and avoid special characters
-            'windowsfilenames': True,   # Avoid characters that are problematic on Windows
-        }
-        
-        downloaded_files = []
-        
-        # Custom hook to capture downloaded files
-        def progress_hook(d):
-            if d['status'] == 'finished':
-                file_path = d['filename']
-                print(f"[DEBUG] yt-dlp reported downloaded file: {file_path}")
-                # Only add files that exist and are not temporary/intermediate files
-                if os.path.exists(file_path) and not any(temp in os.path.basename(file_path) for temp in ['.f', '.part', '.temp']):
-                    downloaded_files.append(file_path)
-                    print(f"[DEBUG] Final file confirmed and added: {file_path}")
-                else:
-                    print(f"[DEBUG] Skipping intermediate/temp file: {file_path}")
-        
-        ydl_opts['progress_hooks'] = [progress_hook]
-        
-        # Download the video
-        print(f"[DEBUG] Starting download from: {url}")
-        print(f"[DEBUG] Output directory: {output_dir}")
-        print(f"[DEBUG] Filename template: {safe_filename}")
-        
-        with yt_dlp.YoutubeDL(ydl_opts) as ydl:
-            ydl.download([url])
-        
-        # Wait a moment for any final processing
-        time.sleep(2)
-        
-        # Additional verification: check for any video files in the output directory
-        if not downloaded_files:
-            print(f"[DEBUG] No files captured by progress hook, scanning directory...")
-            video_extensions = {'.mp4', '.avi', '.mov', '.mkv', '.wmv', '.flv', '.webm', '.m4v'}
-            for file in os.listdir(output_dir):
-                file_path = os.path.join(output_dir, file)
-                if (os.path.splitext(file.lower())[1] in video_extensions and 
-                    os.path.isfile(file_path) and
-                    not any(temp in file for temp in ['.f', '.part', '.temp'])):
-                    # Check if this is a recently created file (within last 120 seconds)
-                    if os.path.getmtime(file_path) > time.time() - 120:
-                        downloaded_files.append(file_path)
-                        print(f"[DEBUG] Found recent video file: {file_path}")
-        
-        if downloaded_files:
-            # Sort by modification time and take the most recent
-            downloaded_files.sort(key=lambda x: os.path.getmtime(x), reverse=True)
-            final_path = downloaded_files[0]
-            file_size = os.path.getsize(final_path)
-            print(f"[DEBUG] Successfully downloaded: {final_path} ({file_size} bytes)")
-            return True, f"Successfully downloaded video ({file_size} bytes)", final_path
-        else:
-            return False, "Download completed but no valid video file was created", None
-            
-    except Exception as e:
-        error_msg = str(e)
-        print(f"[DEBUG] Download exception: {error_msg}")
-        if "Private video" in error_msg:
-            return False, "Video is private or requires authentication", None
-        elif "Video unavailable" in error_msg:
-            return False, "Video is unavailable or has been removed", None
-        elif "Unsupported URL" in error_msg:
-            return False, f"URL not supported by yt-dlp: {url}", None
-        else:
-            return False, f"Download failed: {error_msg}", None
-
-def extract_frames_from_video(video_path, output_folder, frames_per_scene=2, scene_threshold=30.0):
-    """
-    Extract keyframes from a video using scene detection.
-    
-    Args:
-        video_path: Path to the input video file
-        output_folder: Folder to save extracted frames
-        frames_per_scene: Number of frames to extract per detected scene
-        scene_threshold: Threshold for scene detection (lower = more sensitive)
-        
-    Returns:
-        tuple: (success, message, list_of_extracted_frames)
-    """
-    if not VIDEO_PROCESSING_AVAILABLE:
-        print("[WARNING] Scene detection not available, falling back to uniform extraction")
-        return extract_uniform_frames(video_path, output_folder, 20)  # Fallback to uniform extraction
-        
-    try:
-        # Create output folder if it doesn't exist
-        os.makedirs(output_folder, exist_ok=True)
-        
-        print(f"[DEBUG] Initializing video manager for: {video_path}")
-        
-        # Initialize video manager and scene manager for scene detection
-        video_manager = VideoManager([video_path])
-        scene_manager = SceneManager()
-        scene_manager.add_detector(ContentDetector(threshold=scene_threshold))
-        
-        # Detect scenes
-        video_manager.set_duration()
-        video_manager.start()
-        scene_manager.detect_scenes(frame_source=video_manager)
-        scene_list = scene_manager.get_scene_list()
-        video_manager.release()
-        
-        print(f"[VIDEO] Detected {len(scene_list)} scenes in {video_path}")
-        
-        # If no scenes detected, use uniform sampling
-        if not scene_list:
-            print("[WARNING] No scenes detected, falling back to uniform extraction")
-            return extract_uniform_frames(video_path, output_folder, 20)
-        
-        # Extract keyframes from scenes
-        extracted_frames = []
-        
-        # Use OpenCV for frame extraction
-        cap = cv2.VideoCapture(video_path)
-        if not cap.isOpened():
-            print(f"[ERROR] OpenCV could not open video: {video_path}")
-            return False, f"Could not open video file with OpenCV: {video_path}", []
-            
-        total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-        fps = cap.get(cv2.CAP_PROP_FPS)
-        
-        print(f"[DEBUG] Video properties - Total frames: {total_frames}, FPS: {fps}")
-        
-        frame_count = 0
-        for i, scene in enumerate(scene_list):
-            start_time = scene[0].get_seconds()
-            end_time = scene[1].get_seconds()
-            
-            # Extract frames uniformly from this scene
-            scene_duration = end_time - start_time
-            if scene_duration > 0:
-                for j in range(frames_per_scene):
-                    # Calculate frame position within scene
-                    time_offset = (j + 0.5) * scene_duration / frames_per_scene
-                    frame_time = start_time + time_offset
-                    frame_number = int(frame_time * fps)
-                    
-                    # Extract frame
-                    cap.set(cv2.CAP_PROP_POS_FRAMES, frame_number)
-                    ret, frame = cap.read()
-                    
-                    if ret:
-                        # Convert BGR to RGB
-                        frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-                        
-                        # Save frame
-                        video_name = os.path.splitext(os.path.basename(video_path))[0]
-                        frame_filename = f"{video_name}_scene{i:03d}_frame{j:03d}.jpg"
-                        frame_path = os.path.join(output_folder, frame_filename)
-                        
-                        # Convert to PIL and save
-                        pil_image = Image.fromarray(frame_rgb)
-                        pil_image.save(frame_path, 'JPEG', quality=95)
-                        
-                        extracted_frames.append(frame_path)
-                        frame_count += 1
-                        
-                        print(f"[VIDEO] Extracted frame {frame_count}/{frames_per_scene * len(scene_list)}: {frame_filename}")
-        
-        cap.release()
-        
-        return True, f"Successfully extracted {len(extracted_frames)} keyframes from {len(scene_list)} scenes", extracted_frames
-        
-    except Exception as e:
-        print(f"[ERROR] Video processing failed: {e}")
-        import traceback
-        traceback.print_exc()
-        return False, f"Error processing video: {str(e)}", []
-
-def extract_uniform_frames(video_path, output_folder, num_frames=20):
-    """
-    Extract frames uniformly distributed across the video duration.
-    
-    Args:
-        video_path: Path to the input video file
-        output_folder: Folder to save extracted frames
-        num_frames: Maximum number of frames to extract
-    
-    Returns:
-        tuple: (success, message, extracted_frames_list)
-    """
-    try:
-        os.makedirs(output_folder, exist_ok=True)
-        
-        cap = cv2.VideoCapture(video_path)
-        if not cap.isOpened():
-            return False, f"Could not open video file: {video_path}", []
-            
-        total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-        fps = cap.get(cv2.CAP_PROP_FPS)
-        duration = total_frames / fps
-        
-        print(f"[VIDEO] Video has {total_frames} frames, {duration:.2f} seconds duration")
-        
-        extracted_frames = []
-        frame_interval = max(1, total_frames // num_frames)
-        
-        for i in range(0, total_frames, frame_interval):
-            if len(extracted_frames) >= num_frames:
-                break
-                
-            cap.set(cv2.CAP_PROP_POS_FRAMES, i)
-            ret, frame = cap.read()
-            
-            if ret:
-                # Convert BGR to RGB
-                frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-                
-                # Save frame
-                video_name = os.path.splitext(os.path.basename(video_path))[0]
-                frame_filename = f"{video_name}_uniform_{len(extracted_frames):03d}.jpg"
-                frame_path = os.path.join(output_folder, frame_filename)
-                
-                # Convert to PIL and save
-                pil_image = Image.fromarray(frame_rgb)
-                pil_image.save(frame_path, 'JPEG', quality=95)
-                
-                extracted_frames.append(frame_path)
-                print(f"[VIDEO] Extracted uniform frame {len(extracted_frames)}/{num_frames}: {frame_filename}")
-        
-        cap.release()
-        
-        return True, f"Successfully extracted {len(extracted_frames)} frames uniformly", extracted_frames
-        
-    except Exception as e:
-        print(f"[ERROR] Uniform frame extraction failed: {e}")
-        import traceback
-        traceback.print_exc()
-        return False, f"Error extracting uniform frames: {str(e)}", []
-
-def extract_frames_with_progress(urls_text, output_folder, frames_per_scene, scene_threshold, max_quality):
-    """Process videos from URLs with progress updates for Gradio"""
-    if not YT_DLP_AVAILABLE:
-        return "❌ yt-dlp not available. Please install it: pip install yt-dlp"
-    
-    if not VIDEO_PROCESSING_AVAILABLE:
-        return "⚠️ Scene detection not available, will fall back to uniform extraction"
-    
-    # Parse URLs from input text (either comma-separated or one per line)
-    if ',' in urls_text:
-        urls = [url.strip() for url in urls_text.split(',') if url.strip()]
-    else:
-        urls = [url.strip() for url in urls_text.split('\n') if url.strip()]
-    
-    if not urls:
-        return "❌ No URLs provided"
-    
-    # Filter out invalid URLs
-    valid_urls = [url for url in urls if is_supported_video_url(url)]
-    invalid_urls = [url for url in urls if not is_supported_video_url(url)]
-    
-    status_messages = []
-    
-    if invalid_urls:
-        status_messages.append(f"⚠️ Skipping {len(invalid_urls)} invalid/unsupported URLs: {', '.join(invalid_urls)}")
-    
-    if not valid_urls:
-        status_messages.append("❌ No valid video URLs found")
-        return "\n".join(status_messages)
-    
-    status_messages.append(f"🔗 Found {len(valid_urls)} valid video URLs to process")
-    
-    # Create temporary directory for downloads
-    temp_download_dir = tempfile.mkdtemp(prefix="reverso_downloads_")
-    
-    try:
-        total_extracted = 0
-        successful_videos = 0
-        downloaded_files = []
-        
-        # Download all videos first
-        status_messages.append(f"📥 Starting downloads to temporary directory...")
-        
-        for i, url in enumerate(valid_urls):
-            status_messages.append(f"📥 Downloading video {i+1}/{len(valid_urls)}: {url}")
-            
-            try:
-                success, message, downloaded_file = download_video_from_url(url, temp_download_dir, max_quality)
-                
-                if success and downloaded_file:
-                    downloaded_files.append(downloaded_file)
-                    status_messages.append(f"✅ Downloaded: {os.path.basename(downloaded_file)}")
-                else:
-                    status_messages.append(f"❌ Failed to download {url}: {message}")
-                    
-            except Exception as e:
-                status_messages.append(f"❌ Error downloading {url}: {str(e)}")
-        
-        if not downloaded_files:
-            status_messages.append("❌ No videos were successfully downloaded")
-            return "\n".join(status_messages)
-        
-        status_messages.append(f"🎬 Successfully downloaded {len(downloaded_files)} videos. Starting frame extraction...")
-        
-        # Now process the downloaded videos
-        for i, video_path in enumerate(downloaded_files):
-            video_name = os.path.basename(video_path)
-            status_messages.append(f"🎬 Processing video {i+1}/{len(downloaded_files)}: {video_name}")
-            
-            # Create subfolder for this video's frames
-            video_output_folder = os.path.join(output_folder, os.path.splitext(video_name)[0])
-            
-            try:
-                success, message, extracted_frames = extract_frames_from_video(
-                    video_path, video_output_folder, frames_per_scene, scene_threshold
-                )
-                
-                if success:
-                    total_extracted += len(extracted_frames)
-                    successful_videos += 1
-                    status_messages.append(f"✅ {video_name}: {message}")
-                else:
-                    status_messages.append(f"❌ {video_name}: {message}")
-                    
-            except Exception as e:
-                status_messages.append(f"❌ {video_name}: Error - {str(e)}")
-        
-        status_messages.append(f"🎉 Processing complete! Successfully processed {successful_videos}/{len(downloaded_files)} videos")
-        status_messages.append(f"📊 Total frames extracted: {total_extracted}")
-        status_messages.append(f"💾 Frames saved to: {output_folder}")
-        
-    finally:
-        # Cleanup temporary directory
-        try:
-            shutil.rmtree(temp_download_dir)
-            status_messages.append(f"🧹 Cleaned up temporary downloads")
-        except Exception as e:
-            status_messages.append(f"⚠️ Warning: Could not clean up temporary directory: {str(e)}")
-    
-    return "\n".join(status_messages)
-
-def process_local_videos_with_progress(input_folder, output_folder, frames_per_scene, scene_threshold):
-    """Process local video files with progress updates for Gradio"""
-    if not VIDEO_PROCESSING_AVAILABLE:
-        return "⚠️ Scene detection not available, will fall back to uniform extraction"
-    
-    # Supported video extensions
-    video_extensions = {'.mp4', '.avi', '.mov', '.mkv', '.wmv', '.flv', '.webm', '.m4v'}
-    
-    # Find all video files
-    video_files = []
-    for file in os.listdir(input_folder):
-        if os.path.splitext(file.lower())[1] in video_extensions:
-            video_files.append(os.path.join(input_folder, file))
-    
-    if not video_files:
-        return "❌ No video files found in the specified folder"
-    
-    status_messages = []
-    status_messages.append(f"📹 Found {len(video_files)} video files to process")
-    
-    total_extracted = 0
-    successful_videos = 0
-    
-    for i, video_path in enumerate(video_files):
-        video_name = os.path.basename(video_path)
-        status_messages.append(f"🎬 Processing video {i+1}/{len(video_files)}: {video_name}")
-        
-        # Create subfolder for this video's frames
-        video_output_folder = os.path.join(output_folder, os.path.splitext(video_name)[0])
-        
-        try:
-            success, message, extracted_frames = extract_frames_from_video(
-                video_path, video_output_folder, frames_per_scene, scene_threshold
-            )
-            
-            if success:
-                total_extracted += len(extracted_frames)
-                successful_videos += 1
-                status_messages.append(f"✅ {video_name}: {message}")
-            else:
-                status_messages.append(f"❌ {video_name}: {message}")
-                
-        except Exception as e:
-            status_messages.append(f"❌ {video_name}: Error - {str(e)}")
-    
-    status_messages.append(f"🎉 Processing complete! Successfully processed {successful_videos}/{len(video_files)} videos")
-    status_messages.append(f"📊 Total frames extracted: {total_extracted}")
-    status_messages.append(f"💾 Frames saved to: {output_folder}")
-    
-    return "\n".join(status_messages)
-
-# =============================================================================
-# CORE SYSTEM: Simplified PE + GroundedSAM Pipeline
-# =============================================================================
-
-class SimpleReverso:
-    """Simplified visual investigation system"""
-    
-    def __init__(self):
-        print("🚀 Initializing Simple Revers-o...")
-        
-        # Setup device
-        self.device = self.setup_device()
-        
-        # Load optimized PE model (L14-336, layer 24)
-        self.pe_model, self.preprocess = self.load_pe_model()
-        
-        # Initialize GroundedSAM
-        self.grounded_sam = None
-        
-        # Vector database
-        self.vector_db = None
-        self.current_database = None
-        
-        # State
-        self.detected_regions = []
-        self.region_embeddings = []
-        
-        print("✅ Simple Revers-o ready!")
-    
-    def list_databases(self):
-        """List all available databases"""
-        db_path = "./simple_reverso_db"
-        if not os.path.exists(db_path):
-            return []
-        
-        databases = []
-        for name in os.listdir(db_path):
-            db_dir = os.path.join(db_path, name)
-            if os.path.isdir(db_dir):
-                # Check if it's a valid database by looking for meta.json
-                if os.path.exists(os.path.join(db_dir, "meta.json")):
-                    databases.append(name)
-        return databases
-    
-    def load_database(self, database_name):
-        """Load an existing database"""
-        if not database_name:
-            return "❌ Please provide a database name"
-        
-        db_path = f"./simple_reverso_db/{database_name}"
-        if not os.path.exists(db_path):
-            return f"❌ Database not found: {database_name}"
-        
-        try:
-            client = QdrantClient(path=db_path)
-            collection_name = f"simple_reverso_{database_name}"
-            
-            # Verify collection exists
-            collections = client.get_collections().collections
-            if not any(c.name == collection_name for c in collections):
-                return f"❌ Collection not found in database: {database_name}"
-            
-            # Store database info
-            self.vector_db = client
-            self.current_database = collection_name
-            
-            return f"✅ Loaded database: {database_name}"
-            
-        except Exception as e:
-            return f"❌ Error loading database: {str(e)}"
-    
-    def delete_database(self, database_name):
-        """Delete a database"""
-        if not database_name:
-            return "❌ Please provide a database name"
-        
-        db_path = f"./simple_reverso_db/{database_name}"
-        if not os.path.exists(db_path):
-            return f"❌ Database not found: {database_name}"
-        
-        try:
-            # Remove the database directory
-            shutil.rmtree(db_path)
-            return f"✅ Deleted database: {database_name}"
-        except Exception as e:
-            return f"❌ Error deleting database: {str(e)}"
-    
-    def unlock_database(self, database_name):
-        """Remove the lock file from a database"""
-        if not database_name:
-            return "❌ Please provide a database name"
-        
-        db_path = f"./simple_reverso_db/{database_name}"
-        if not os.path.exists(db_path):
-            return f"❌ Database not found: {database_name}"
-        
-        lock_file = os.path.join(db_path, ".lock")
-        if os.path.exists(lock_file):
-            try:
-                os.remove(lock_file)
-                return f"✅ Removed lock file from database: {database_name}"
-            except Exception as e:
-                return f"❌ Error removing lock file: {str(e)}"
-        else:
-            return f"ℹ️ No lock file found for database: {database_name}"
-    
-    def setup_device(self):
-        """Setup optimal compute device"""
-        if torch.backends.mps.is_available():
-            device = torch.device("mps")
-            print(f"🔥 Using MPS: {device}")
-        elif torch.cuda.is_available():
-            device = torch.device("cuda")
-            print(f"🔥 Using CUDA: {device}")
-        else:
-            device = torch.device("cpu")
-            print(f"💻 Using CPU: {device}")
-        return device
-    
-    def load_pe_model(self):
-        """Load PE-Core-L14-336 - optimal for investigation tasks"""
-        print("📚 Loading PE-Core-L14-336 (optimal for investigation)...")
-        
-        available_configs = pe.CLIP.available_configs()
-        print(f"Available PE configs: {available_configs}")
-        
-        # Target PE-Core-L14-336 (1B params, layer 24 optimal)
-        target_model = "PE-Core-L14-336"
-        
-        if target_model in available_configs:
-            try:
-                pe_model = pe.CLIP.from_config(target_model, pretrained=True)
-                print(f"✅ Loaded {target_model}")
-            except Exception as e:
-                print(f"❌ Failed to load {target_model}: {e}")
-                # Fallback to first available
-                pe_model = pe.CLIP.from_config(available_configs[0], pretrained=True)
-                print(f"🔄 Using fallback: {available_configs[0]}")
-        else:
-            # Use first available
-            pe_model = pe.CLIP.from_config(available_configs[0], pretrained=True)
-            print(f"🔄 Using available: {available_configs[0]}")
-        
-        # Move to device and optimize
-        pe_model = pe_model.to(self.device)
-        if self.device.type == 'cuda':
-            pe_model = pe_model.half()  # Mixed precision
-            print("⚡ Mixed precision enabled")
-        
-        # Preprocessing for 336px (L14 optimal)
-        preprocess = transforms.get_image_transform(336)
-        
-        print(f"🎯 PE model ready - using layer 24 (research optimal)")
-        return pe_model, preprocess
-    
-    def init_grounded_sam(self, text_prompt):
-        """Initialize GroundedSAM with text prompt"""
-        if self.grounded_sam is None:
-            # Parse prompts
-            prompts = [p.strip() for p in text_prompt.split('.') if p.strip()]
-            if not prompts:
-                prompts = ["object"]
-            
-            # Create ontology
-            ontology_dict = {prompt: prompt for prompt in prompts}
-            ontology = CaptionOntology(ontology_dict)
-            
-            # Initialize GroundedSAM with correct parameters
-            self.grounded_sam = GroundedSAM(
-                ontology=ontology,
-                box_threshold=0.35,
-                text_threshold=0.25
-            )
-            print(f"🎯 GroundedSAM ready with prompts: {prompts}")
-            print(f"[DEBUG] GroundedSAM version: {self.grounded_sam.__version__ if hasattr(self.grounded_sam, '__version__') else 'unknown'}")
-            print(f"[DEBUG] GroundedSAM config: {self.grounded_sam.__dict__}")
-            
-            # Print available methods and attributes
-            print(f"[DEBUG] Available methods: {[m for m in dir(self.grounded_sam) if not m.startswith('_')]}")
-            
-            # Print predict method signature
-            if hasattr(self.grounded_sam, 'predict'):
-                import inspect
-                print(f"[DEBUG] Predict method signature: {inspect.signature(self.grounded_sam.predict)}")
-    
-    def detect_regions(self, image, text_prompt="person . car . building"):
-        """Detect regions using GroundedSAM"""
-        print(f"🔍 Detecting regions with prompt: '{text_prompt}'")
-        
-        # Initialize GroundedSAM if needed
-        self.init_grounded_sam(text_prompt)
-        
-        # Save image temporarily for GroundedSAM
-        temp_path = f"/tmp/temp_image_{uuid.uuid4().hex[:8]}.jpg"
-        if isinstance(image, np.ndarray):
-            image_pil = Image.fromarray(image)
-        else:
-            image_pil = image
-        
-        image_pil.save(temp_path)
-        
-        try:
-            # Detect with GroundedSAM
-            detections = self.grounded_sam.predict(temp_path)
-            print(f"✅ Found {len(detections)} regions")
-            print(f"[DEBUG] Detection type: {type(detections)}")
-            print(f"[DEBUG] Detection attributes: {dir(detections)}")
-            
-            # Convert detections to supervision format with masks
-            from supervision.detection.core import Detections
-            
-            # Extract boxes and masks
-            boxes = detections.xyxy
-            class_ids = detections.class_id
-            confidences = detections.confidence
-            
-            # Get masks from the predictions
-            masks = []
-            if hasattr(detections, 'masks'):
-                masks = detections.masks
-            elif hasattr(detections, 'mask'):
-                masks = detections.mask
-            elif hasattr(detections, 'data') and 'masks' in detections.data:
-                masks = detections.data['masks']
-            
-            print(f"[DEBUG] Boxes shape: {boxes.shape if hasattr(boxes, 'shape') else 'no shape'}")
-            print(f"[DEBUG] Masks type: {type(masks)}")
-            print(f"[DEBUG] Masks shape: {masks.shape if hasattr(masks, 'shape') else 'no shape'}")
-            print(f"[DEBUG] Class IDs: {class_ids}")
-            print(f"[DEBUG] Confidences: {confidences}")
-            
-            # Convert masks to numpy if they're tensors
-            if isinstance(masks, torch.Tensor):
-                masks = masks.detach().cpu().numpy()
-                print(f"[DEBUG] Converted masks to numpy array")
-            
-            # Create supervision detections with masks
-            self.detected_regions = Detections(
-                xyxy=boxes,
-                mask=masks,
-                confidence=confidences,
-                class_id=class_ids
-            )
-            
-            # Debug print the created detections
-            print(f"[DEBUG] Created detections type: {type(self.detected_regions)}")
-            print(f"[DEBUG] Created detections attributes: {dir(self.detected_regions)}")
-            print(f"[DEBUG] Created detections mask attribute: {hasattr(self.detected_regions, 'mask')}")
-            if hasattr(self.detected_regions, 'mask'):
-                print(f"[DEBUG] Created detections mask shape: {self.detected_regions.mask.shape if hasattr(self.detected_regions.mask, 'shape') else 'no shape'}")
-            
-            return len(detections)
-            
-        finally:
-            # Cleanup
-            if os.path.exists(temp_path):
-                os.remove(temp_path)
-    
-    def extract_embeddings(self, image):
-        """Extract PE embeddings from detected regions"""
-        if not self.detected_regions or len(self.detected_regions) == 0:
-            print("❌ No regions detected")
-            return [], []
-        
-        print(f"🧠 Extracting PE embeddings from {len(self.detected_regions)} regions...")
-        
-        # Convert image to tensor
-        if isinstance(image, np.ndarray):
-            image_pil = Image.fromarray(image)
-        else:
-            image_pil = image
-        
-        image_tensor = self.preprocess(image_pil).unsqueeze(0).to(self.device)
-        
-        embeddings = []
-        metadata = []
-        
-        with torch.no_grad():
-            # Extract features using standard PE encoding
-            features = self.pe_model.encode_image(image_tensor)
-            print(f"✅ Using standard PE encoding (still very effective)")
-            print(f"[DEBUG] Raw features shape: {features.shape}")
-            
-            # Handle different feature shapes
-            if len(features.shape) == 3:  # [1, num_tokens, dim]
-                B, N, D = features.shape
-                print(f"[DEBUG] Token-based features: {B} batch, {N} tokens, {D} dimensions")
-                
-                # For token-based features, we'll use a simpler approach
-                # Just use the global image embedding for all regions
-                global_embedding = features.mean(dim=1)  # [1, D] - average all tokens
-                print(f"[DEBUG] Using global embedding approach due to token format")
-                
-            elif len(features.shape) == 2:  # [1, dim] - already a global embedding
-                global_embedding = features
-                print(f"[DEBUG] Global embedding shape: {global_embedding.shape}")
-                
-            else:
-                print(f"[ERROR] Unexpected feature shape: {features.shape}")
-                return [], []
-            
-            # Get the class names from the ontology
-            try:
-                ontology_class_names = self.grounded_sam.ontology.classes()
-                print(f"[DEBUG] Ontology class names: {ontology_class_names}")
-            except Exception as e:
-                print(f"[WARNING] Could not get ontology classes: {e}")
-                # Fallback to extracting from the ontology dict
-                try:
-                    ontology_class_names = list(self.grounded_sam.ontology.prompts_to_classes.keys())
-                except:
-                    ontology_class_names = ["object"]  # Final fallback
-            
-            # Process each detected region
-            for i in range(min(len(self.detected_regions), 10)):
-                try:
-                    # Get the mask directly from the detections object
-                    if hasattr(self.detected_regions, 'mask') and self.detected_regions.mask is not None:
-                        mask_np = self.detected_regions.mask[i]
-                    else:
-                        print(f"❌ No mask available for detection {i}, skipping")
-                        continue
-                    
-                    # Get confidence
-                    raw_confidence = 0.0
-                    if hasattr(self.detected_regions, 'confidence') and self.detected_regions.confidence is not None:
-                        if i < len(self.detected_regions.confidence):
-                            raw_confidence = float(self.detected_regions.confidence[i])
-                    
-                    # Get detected class
-                    detected_class = "object"  # default
-                    if hasattr(self.detected_regions, 'class_id') and self.detected_regions.class_id is not None:
-                        if i < len(self.detected_regions.class_id):
-                            class_id = int(self.detected_regions.class_id[i])
-                            if 0 <= class_id < len(ontology_class_names):
-                                detected_class = ontology_class_names[class_id]
-                    
-                    print(f"[STATUS] Processing detection {i+1}: {detected_class} (Confidence: {raw_confidence:.3f})")
-                    
-                    # Ensure mask is binary uint8
-                    if mask_np.dtype == bool:
-                        mask_processed = mask_np.astype(np.uint8)
-                    elif np.issubdtype(mask_np.dtype, np.floating):
-                        mask_processed = (mask_np > 0.5).astype(np.uint8)
-                    else:
-                        mask_processed = mask_np.astype(np.uint8)
-                    
-                    # Check if mask is empty
-                    if np.sum(mask_processed) == 0:
-                        print(f"⚠️ Empty mask for region {i}, skipping")
-                        continue
-                    
-                    # For now, use the global image embedding for each region
-                    # This is simpler and still effective for similarity search
-                    region_embedding = global_embedding[0]  # Remove batch dimension
-                    
-                    # Normalize
-                    region_embedding = region_embedding / region_embedding.norm()
-                    
-                    # Store
-                    embeddings.append(region_embedding.cpu())
-                    
-                    # Create metadata - get bounding box from mask
-                    y_indices, x_indices = np.where(mask_processed)
-                    bbox = [int(x_indices.min()), int(y_indices.min()), 
-                           int(x_indices.max()), int(y_indices.max())]
-                    
-                    meta = {
-                        "region_id": str(uuid.uuid4()),
-                        "bbox": bbox,
-                        "area_ratio": float(np.sum(mask_processed) / mask_processed.size),
-                        "detection_index": i,
-                        "confidence": raw_confidence,
-                        "detected_class": detected_class
-                    }
-                    metadata.append(meta)
-                    
-                    print(f"✅ Extracted embedding for region {i}")
-                    
-                except Exception as e:
-                    print(f"❌ Error processing region {i}: {e}")
-                    import traceback
-                    traceback.print_exc()
-                    continue
-        
-        self.region_embeddings = embeddings
-        print(f"🎯 Extracted {len(embeddings)} region embeddings")
-        return embeddings, metadata
-    
-    def process_image_direct_pe(self, image):
-        """Process image directly with PE, without GroundedSAM"""
-        print("🧠 Processing image directly with PE...")
-        
-        # Convert image to tensor
-        if isinstance(image, np.ndarray):
-            image_pil = Image.fromarray(image)
-        else:
-            image_pil = image
-        
-        image_tensor = self.preprocess(image_pil).unsqueeze(0).to(self.device)
-        
-        with torch.no_grad():
-            # Extract features using standard PE encoding
-            features = self.pe_model.encode_image(image_tensor)
-            
-            # Handle different feature shapes
-            if len(features.shape) == 3:  # [1, num_tokens, dim]
-                # Use global embedding (average of tokens)
-                embedding = features.mean(dim=1)[0]  # Remove batch dimension
-            elif len(features.shape) == 2:  # [1, dim]
-                embedding = features[0]  # Remove batch dimension
-            else:
-                raise ValueError(f"Unexpected feature shape: {features.shape}")
-            
-            # Normalize
-            embedding = embedding / embedding.norm()
-            
-            # Store as a single region
-            self.region_embeddings = [embedding.cpu()]
-            
-            # Create metadata for the whole image
-            meta = {
-                "region_id": str(uuid.uuid4()),
-                "bbox": [0, 0, image_pil.width, image_pil.height],
-                "area_ratio": 1.0,
-                "detection_index": 0,
-                "confidence": 1.0,
-                "detected_class": "full_image"
-            }
-            
-            print("✅ Extracted global image embedding")
-            return [embedding.cpu()], [meta]
-
-    def create_database(self, folder_path, database_name, text_prompt="person . car . building", use_direct_pe=False):
-        """Create searchable database from image folder"""
-        status_messages = []
-        
-        def log_status(message):
-            status_messages.append(message)
-            return "\n".join(status_messages)
-        
-        log_status(f"📁 Creating database '{database_name}' from {folder_path}")
-        
-        # Get image files
-        image_extensions = ['.jpg', '.jpeg', '.png', '.bmp', '.tiff']
-        image_files = []
-        
-        for file in os.listdir(folder_path):
-            if any(file.lower().endswith(ext) for ext in image_extensions):
-                image_files.append(os.path.join(folder_path, file))
-        
-        if not image_files:
-            return log_status(f"❌ No images found in {folder_path}")
-        
-        log_status(f"📊 Found {len(image_files)} images to process")
-        log_status(f"🔧 Processing mode: {'Direct PE' if use_direct_pe else 'GroundedSAM + PE'}")
-        
-        # Initialize vector database
-        db_path = f"./simple_reverso_db/{database_name}"
-        os.makedirs(db_path, exist_ok=True)
-        log_status(f"📂 Created database directory: {db_path}")
-        
-        client = QdrantClient(path=db_path)
-        
-        # Process images
-        all_embeddings = []
-        all_metadata = []
-        processed = 0
-        failed = 0
-        
-        for i, image_path in enumerate(image_files):
-            try:
-                filename = os.path.basename(image_path)
-                log_status(f"🔄 Processing {i+1}/{len(image_files)}: {filename}")
-                
-                # Load and process image
-                image = Image.open(image_path).convert("RGB")
-                
-                if use_direct_pe:
-                    # Process directly with PE
-                    embeddings, metadata = self.process_image_direct_pe(image)
-                    log_status(f"✅ Extracted global embedding for {filename}")
-                else:
-                    # Detect regions first
-                    num_regions = self.detect_regions(image, text_prompt)
-                    if num_regions > 0:
-                        embeddings, metadata = self.extract_embeddings(image)
-                        log_status(f"✅ Found {num_regions} regions in {filename}")
-                    else:
-                        log_status(f"⚠️ No regions found in {filename}, skipping")
-                        failed += 1
-                        continue
-                    
-                    # Add source info to metadata
-                    for meta in metadata:
-                        meta["image_source"] = image_path
-                    meta["filename"] = filename
-                    
-                    all_embeddings.extend(embeddings)
-                    all_metadata.extend(metadata)
-                    processed += 1
-                
-            except Exception as e:
-                log_status(f"❌ Error processing {filename}: {str(e)}")
-                failed += 1
-                continue
-        
-        if not all_embeddings:
-            return log_status(f"❌ No regions found in any images")
-        
-        # Create collection
-        vector_dim = all_embeddings[0].shape[0]
-        collection_name = f"simple_reverso_{database_name}"
-        
-        try:
-            client.create_collection(
-                collection_name=collection_name,
-                vectors_config=models.VectorParams(
-                    size=vector_dim,
-                    distance=models.Distance.COSINE
-                )
-            )
-            log_status(f"📦 Created collection: {collection_name}")
-        except Exception:
-            log_status(f"🔄 Collection {collection_name} already exists")
-        
-        # Insert embeddings
-        log_status(f"💾 Storing {len(all_embeddings)} embeddings...")
-        points = []
-        for i, (emb, meta) in enumerate(zip(all_embeddings, all_metadata)):
-            points.append(models.PointStruct(
-                id=meta["region_id"],
-                vector=emb.numpy().tolist(),
-                payload=meta
-            ))
-        
-        client.upsert(collection_name=collection_name, points=points)
-        
-        # Store database info
-        self.vector_db = client
-        self.current_database = collection_name
-        
-        # Final summary
-        log_status("\n📊 Final Summary:")
-        log_status(f"✅ Successfully processed: {processed} images")
-        if failed > 0:
-            log_status(f"⚠️ Failed to process: {failed} images")
-        log_status(f"🔍 Total embeddings stored: {len(all_embeddings)}")
-        log_status(f"🎯 Database ready for searching!")
-        
-        return "\n".join(status_messages)
-    
-    def search_similar(self, similarity_threshold=0.7, max_results=5):
-        """Search for similar regions in database"""
-        if not self.region_embeddings:
-            return "❌ No regions detected. Please detect regions first.", []
-        
-        if not self.vector_db or not self.current_database:
-            return "❌ No database loaded. Please create a database first.", []
-        
-        print(f"🔍 Searching for similar regions (threshold={similarity_threshold})")
-        
-        # Use first detected region for search
-        query_embedding = self.region_embeddings[0]
-        
-        # Search
-        search_results = self.vector_db.search(
-            collection_name=self.current_database,
-            query_vector=query_embedding.numpy().tolist(),
-            limit=max_results,
-            score_threshold=similarity_threshold
-        )
-        
-        if not search_results:
-            return f"❌ No similar regions found above threshold {similarity_threshold}", []
-        
-        # Format results
-        results_text = f"🎯 Found {len(search_results)} similar regions:\n\n"
-        similar_images = []
-        
-        for i, result in enumerate(search_results):
-            filename = result.payload.get("filename", "Unknown")
-            score = result.score
-            bbox = result.payload.get("bbox", [0,0,0,0])
-            image_path = result.payload.get("image_source", "")
-            
-            results_text += f"{i+1}. {filename} (similarity: {score:.3f})\n"
-            results_text += f"   📍 Bounding box: {bbox}\n\n"
-        
-            # Load and process the similar image
-            if os.path.exists(image_path):
-                try:
-                    # Load the image
-                    img = Image.open(image_path).convert("RGB")
-                    
-                    # If we have a bounding box, crop to that region
-                    if bbox != [0, 0, img.width, img.height]:  # Not the full image
-                        x1, y1, x2, y2 = bbox
-                        img = img.crop((x1, y1, x2, y2))
-                    
-                    # Add similarity score as text overlay
-                    img_with_text = img.copy()
-                    draw = ImageDraw.Draw(img_with_text)
-                    # Use a larger font size and add a background for better visibility
-                    font_size = max(20, int(img.height * 0.05))  # Scale font size with image height
-                    try:
-                        from PIL import ImageFont
-                        font = ImageFont.truetype("Arial", font_size)
-                    except:
-                        font = ImageFont.load_default()
-                    
-                    # Add text with background
-                    text = f"Score: {score:.3f}"
-                    text_bbox = draw.textbbox((10, 10), text, font=font)
-                    draw.rectangle([text_bbox[0]-5, text_bbox[1]-5, text_bbox[2]+5, text_bbox[3]+5], 
-                                 fill='black', outline='white')
-                    draw.text((10, 10), text, fill='white', font=font)
-                    
-                    # Resize if too large
-                    max_size = (800, 800)
-                    if img_with_text.width > max_size[0] or img_with_text.height > max_size[1]:
-                        img_with_text.thumbnail(max_size, Image.Resampling.LANCZOS)
-                    
-                    # Store both the image and its score
-                    similar_images.append((img_with_text, score))
-                    print(f"✅ Loaded image {i+1}: {image_path}")
-                except Exception as e:
-                    print(f"❌ Error loading image {image_path}: {e}")
-                    similar_images.append(None)
-            else:
-                print(f"❌ Image not found: {image_path}")
-                similar_images.append(None)
-        
-        print(f"📊 Found {len(similar_images)} images to display")
-        return results_text, similar_images
-    
-    def visualize_detections(self, image, selected_region_index=None):
-        """Create visualization of detected regions, highlighting the selected region if provided"""
-        if not self.detected_regions:
-            return None
-        
-        # Convert image to RGB if it's not already
-        if isinstance(image, np.ndarray):
-            if len(image.shape) == 2:  # Grayscale
-                image = cv2.cvtColor(image, cv2.COLOR_GRAY2RGB)
-            elif image.shape[2] == 4:  # RGBA
-                image = cv2.cvtColor(image, cv2.COLOR_RGBA2RGB)
-        else:
-            image = np.array(image.convert('RGB'))
-
-        print(f"\n[DEBUG] Starting visualization of {len(self.detected_regions)} regions")
-        print(f"[DEBUG] Image shape: {image.shape}")
-        print(f"[DEBUG] Detection type: {type(self.detected_regions)}")
-        
-        # Get masks directly from the detections object
-        masks = self.detected_regions.mask
-        print(f"[DEBUG] Masks type: {type(masks)}")
-        print(f"[DEBUG] Masks shape: {masks.shape if hasattr(masks, 'shape') else 'no shape'}")
-
-        fig, ax = plt.subplots(figsize=(10, 8))
-        ax.imshow(image)
-        
-        for i in range(len(self.detected_regions)):
-            print(f"\n[DEBUG] Processing region {i+1}")
-            
-            if masks is not None and i < len(masks):
-                mask = masks[i]
-                print(f"[DEBUG] Mask {i+1} shape: {mask.shape}")
-                
-                # Ensure mask is binary
-                mask = (mask > 0.5).astype(np.uint8)
-                print(f"[DEBUG] After thresholding, mask unique values: {np.unique(mask)}")
-                print(f"[DEBUG] Mask sum (pixels): {np.sum(mask)}")
-                
-                # Find contours
-                contours, _ = cv2.findContours(mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-                print(f"[DEBUG] Found {len(contours)} contours")
-                
-                # Choose color and linewidth based on selection
-                if selected_region_index is not None and i == selected_region_index:
-                    color = 'lime'
-                    label_facecolor = 'green'
-                    linewidth = 4
-                else:
-                    color = 'red'
-                    label_facecolor = 'red'
-                    linewidth = 2
-                
-                # Draw contours
-                for j, contour in enumerate(contours):
-                    if len(contour) >= 3:
-                        contour = contour.squeeze()
-                        if contour.ndim == 2:
-                            print(f"[DEBUG] Drawing contour {j+1} with {len(contour)} points")
-                            ax.plot(contour[:, 0], contour[:, 1], color=color, linewidth=linewidth)
-                        else:
-                            print(f"[DEBUG] Skipping contour {j+1} - invalid shape: {contour.shape}")
-                
-                # Add label
-                y_coords, x_coords = np.where(mask)
-                if len(x_coords) > 0 and len(y_coords) > 0:
-                    center_x = int(x_coords.mean())
-                    center_y = int(y_coords.mean())
-                    ax.text(center_x, center_y, f"{i+1}", 
-                           color='white', fontsize=12, fontweight='bold',
-                            bbox=dict(boxstyle="round,pad=0.3", facecolor=label_facecolor, alpha=0.7))
-                    print(f"[DEBUG] Added label at ({center_x}, {center_y})")
-            else:
-                print(f"[DEBUG] Region {i+1} has no mask")
-        
-        ax.set_title(f"Detected Regions ({len(self.detected_regions)} found)")
-        ax.axis('off')
-        
-        temp_path = f"/tmp/visualization_{uuid.uuid4().hex[:8]}.png"
-        plt.savefig(temp_path, bbox_inches='tight', dpi=150)
-        plt.close()
-        result_image = Image.open(temp_path)
-        os.remove(temp_path)
-        return result_image
-
-# =============================================================================
-# GRADIO INTERFACE: Simple and Clean
-# =============================================================================
-
-def create_simple_interface():
-    """Create simplified Gradio interface"""
-    
-    # Initialize system
-    reverso = SimpleReverso()
-    
-    # Interface functions
-    def detect_and_extract(image, text_prompt, use_direct_pe=False):
-        """Detect regions and extract embeddings"""
-        if image is None:
-            return None, "❌ Please upload an image", None, None
-        
-        try:
-            if use_direct_pe:
-                # Process directly with PE
-                embeddings, metadata = reverso.process_image_direct_pe(image)
-                result_text = (f"✅ Processed image directly with PE\n"
-                             f"🧠 Extracted global image embedding\n"
-                             f"🎯 Ready to search!")
-                # Create a simple visualization of the full image
-                viz_image = image
-                # Return empty region options for direct PE
-                return viz_image, result_text, gr.update(choices=[], value=None, visible=False), None
-            else:
-                # Detect regions
-                num_regions = reverso.detect_regions(image, text_prompt)
-                
-                if num_regions == 0:
-                    return None, f"❌ No regions found with prompt: '{text_prompt}'", gr.update(choices=[], value=None, visible=False), None
-                
-                # Extract embeddings
-                embeddings, metadata = reverso.extract_embeddings(image)
-                
-                # Create visualization
-                viz_image = reverso.visualize_detections(image)
-                
-                # Create region options for dropdown
-                region_options = []
-                for i, meta in enumerate(metadata):
-                    confidence = meta.get('confidence', 0.0)
-                    detected_class = meta.get('detected_class', 'object')
-                    region_options.append(f"Region {i+1}: {detected_class} (Confidence: {confidence:.3f})")
-                
-                result_text = (f"✅ Found {num_regions} regions\n"
-                              f"🧠 Extracted {len(embeddings)} embeddings\n"
-                              f"🎯 Select a region to search with")
-                
-                return viz_image, result_text, gr.update(choices=region_options, value=region_options[0], visible=True), metadata
-            
-        except Exception as e:
-            return None, f"❌ Error: {str(e)}", gr.update(choices=[], value=None, visible=False), None
-    
-    def build_database(folder_path, db_name, db_prompt, use_direct_pe):
-        """Build searchable database"""
-        if not folder_path or not db_name:
-            return "❌ Please provide folder path and database name"
-        
-        if not os.path.exists(folder_path):
-            return f"❌ Folder not found: {folder_path}"
-        
-        try:
-            result = reverso.create_database(folder_path, db_name, db_prompt, use_direct_pe)
-            return result
-        except Exception as e:
-            return f"❌ Error creating database: {str(e)}"
-    
-    def search_database(similarity_threshold, max_results, state, selected_region=None):
-        """Search for similar regions"""
-        try:
-            if selected_region is not None:
-                # Extract region index from selection (e.g., "Region 1: person" -> 0)
-                try:
-                    region_index = int(selected_region.split()[1].rstrip(':')) - 1
-                    # Use the selected region's embedding
-                    if 0 <= region_index < len(reverso.region_embeddings):
-                        reverso.region_embeddings = [reverso.region_embeddings[region_index]]
-                    else:
-                        print(f"⚠️ Invalid region index: {region_index}")
-                except (ValueError, IndexError) as e:
-                    print(f"⚠️ Error parsing region index: {e}")
-                    # Fall back to first region
-                    if reverso.region_embeddings:
-                        reverso.region_embeddings = [reverso.region_embeddings[0]]
-            
-            result_text, similar_images = reverso.search_similar(similarity_threshold, max_results)
-            
-            # Create a list of image options for the dropdown
-            image_options = [f"Image {i+1} (Score: {score:.3f})" for i, (_, score) in enumerate(similar_images) if similar_images[i] is not None]
-            
-            # Create a gallery of thumbnails
-            thumbnail_gallery = []
-            for i, (img, score) in enumerate(similar_images):
-                if img is not None:
-                    # Create a smaller thumbnail version
-                    thumbnail = img.copy()
-                    thumbnail.thumbnail((200, 200))  # Resize to thumbnail size
-                    thumbnail_gallery.append(thumbnail)
-            
-            # Return the results with the new choices and value
-            return (
-                result_text,
-                gr.update(choices=image_options, value=image_options[0] if image_options else None),
-                similar_images[0][0] if similar_images and similar_images[0] is not None else None,
-                similar_images,  # Store images in state
-                thumbnail_gallery  # Return the gallery value directly
-            )
-        except Exception as e:
-            print(f"❌ Search error: {str(e)}")
-            import traceback
-            traceback.print_exc()
-            return f"❌ Search error: {str(e)}", gr.update(choices=[], value=None), None, [], []  # Return empty list for gallery
-    
-    def update_similar_image(selected_image, state):
-        """Update the displayed similar image based on selection"""
-        if not selected_image or not state:
-            return None
-        try:
-            # Extract the image index from the selection (e.g., "Image 1" -> 0)
-            index = int(selected_image.split()[1]) - 1
-            if 0 <= index < len(state) and state[index] is not None:
-                return state[index][0]  # Return the image from the tuple
-            return None
-        except Exception as e:
-            print(f"❌ Error updating image: {str(e)}")
-            return None
-    
-    def list_available_databases():
-        """List all available databases"""
-        databases = reverso.list_databases()
-        if not databases:
-            return "No databases found"
-        return "\n".join(f"- {db}" for db in databases)
-    
-    def load_selected_database(database_name):
-        """Load a selected database"""
-        return reverso.load_database(database_name)
-    
-    def delete_selected_database(database_name):
-        """Delete a selected database"""
-        return reverso.delete_database(database_name)
-    
-    def unlock_selected_database(database_name):
-        """Remove lock file from a selected database"""
-        return reverso.unlock_database(database_name)
-    
-    def update_region_visualization(selected_region, image):
-        """Update the visualization to highlight the selected region"""
-        if image is None or selected_region is None:
-            return None
-        try:
-            region_index = int(selected_region.split()[1].rstrip(':')) - 1
-            return reverso.visualize_detections(image, selected_region_index=region_index)
-        except Exception as e:
-            print(f"❌ Error updating region visualization: {str(e)}")
-            return None
-    
-    # Create interface
-    with gr.Blocks(title="Simple Revers-o: Visual Investigation Tool") as demo:
-        # Add state variable at the top level
-        similar_images_state = gr.State([])
-        detected_image_state = gr.State(None)
-        
-        gr.Markdown("# 🔍 Simple Revers-o: Visual Investigation Tool")
-        gr.Markdown("**GroundedSAM + PE-Core-L14-336 for visual similarity search**")
-        
-        with gr.Tabs():
-            # Tab 0: Extract Video Frames (New)
-            with gr.TabItem("🎬 Extract Video Frames"):
-                gr.Markdown("## Extract Frames from Videos")
-                gr.Markdown("Extract frames from videos for database creation using scene detection")
-                
-                # Use tabs for different input methods
-                with gr.Tabs():
-                    # Tab for URL-based processing
-                    with gr.TabItem("🔗 From URLs"):
-                        if YT_DLP_AVAILABLE:
-                            with gr.Row():
-                                with gr.Column():
-                                    video_urls = gr.TextArea(
-                                        label="Video URLs",
-                                        placeholder="https://www.youtube.com/watch?v=..., https://twitter.com/..., etc.\n(comma-separated or one URL per line)",
-                                        info="Enter video URLs from YouTube, Twitter, Facebook, Instagram, TikTok, etc."
-                                    )
-                                    url_output_folder = gr.Textbox(
-                                        label="Output Folder Path",
-                                        placeholder="/path/to/save/frames",
-                                        info="Folder where extracted frames will be saved"
-                                    )
-                                    with gr.Row():
-                                        frames_per_scene = gr.Slider(
-                                            minimum=1, maximum=10, value=2, step=1,
-                                            label="Frames per Scene",
-                                            info="Number of frames to extract from each detected scene"
-                                        )
-                                        scene_threshold = gr.Slider(
-                                            minimum=10, maximum=60, value=30, step=5,
-                                            label="Scene Detection Threshold",
-                                            info="Lower values detect more scene changes (more sensitive)"
-                                        )
-                                    max_quality = gr.Dropdown(
-                                        choices=["360p", "480p", "720p", "1080p", "best"],
-                                        value="720p",
-                                        label="Max Video Quality",
-                                        info="Higher quality uses more bandwidth and storage"
-                                    )
-                                    url_extract_btn = gr.Button("🎬 Extract Frames from URLs", variant="primary")
-                                    
-                                with gr.Column():
-                                    url_extraction_status = gr.Markdown("URL processing status will appear here")
-                            
-                            # Connect the button to the processing function
-                            url_extract_btn.click(
-                                extract_frames_with_progress,
-                                inputs=[video_urls, url_output_folder, frames_per_scene, scene_threshold, max_quality],
-                                outputs=[url_extraction_status]
-                            )
-                        else:
-                            gr.Markdown("⚠️ **URL video downloading not available.**")
-                            gr.Markdown("Please install yt-dlp to enable URL downloads:")
-                            gr.Code("pip install yt-dlp", language="bash")
-                    
-                    # Tab for local video processing
-                    with gr.TabItem("📁 From Local Files"):
-                        with gr.Row():
-                            with gr.Column():
-                                local_input_folder = gr.Textbox(
-                                    label="Video Folder Path",
-                                    placeholder="/path/to/videos",
-                                    info="Folder containing video files (.mp4, .avi, .mov, etc.)"
-                                )
-                                local_output_folder = gr.Textbox(
-                                    label="Output Folder Path",
-                                    placeholder="/path/to/save/frames",
-                                    info="Folder where extracted frames will be saved"
-                                )
-                                with gr.Row():
-                                    local_frames_per_scene = gr.Slider(
-                                        minimum=1, maximum=10, value=2, step=1,
-                                        label="Frames per Scene",
-                                        info="Number of frames to extract from each detected scene"
-                                    )
-                                    local_scene_threshold = gr.Slider(
-                                        minimum=10, maximum=60, value=30, step=5,
-                                        label="Scene Detection Threshold",
-                                        info="Lower values detect more scene changes (more sensitive)"
-                                    )
-                                local_extract_btn = gr.Button("🎬 Extract Frames from Local Videos", variant="primary")
-                                
-                            with gr.Column():
-                                local_extraction_status = gr.Markdown("Local video processing status will appear here")
-                        
-                        # Connect the button to the processing function
-                        local_extract_btn.click(
-                            process_local_videos_with_progress,
-                            inputs=[local_input_folder, local_output_folder, local_frames_per_scene, local_scene_threshold],
-                            outputs=[local_extraction_status]
-                        )
-            
-            # Tab 1: Create Database
-            with gr.TabItem("🗃️ Create Database"):
-                gr.Markdown("## Build a searchable database from your images")
-                
-                with gr.Row():
-                    with gr.Column():
-                        db_folder = gr.Textbox(
-                            label="📁 Image Folder Path",
-                            placeholder="/path/to/your/images",
-                            info="Folder containing images to process"
-                        )
-                        db_name = gr.Textbox(
-                            label="🏷️ Database Name",
-                            placeholder="my_investigation_db",
-                            info="Name for your database"
-                        )
-                        db_prompt = gr.Textbox(
-                            label="🎯 Detection Prompts",
-                            value="person . car . building",
-                            info="What to look for (period-separated)"
-                        )
-                        use_direct_pe = gr.Checkbox(
-                            label="🔍 Use Direct PE Processing",
-                            value=False,
-                            info="Process images directly with PE (no object detection)"
-                        )
-                        build_btn = gr.Button("🚀 Build Database", variant="primary")
-                    
-                    with gr.Column():
-                        db_status = gr.Textbox(
-                            label="📊 Database Status",
-                            lines=8,
-                            info="Progress and results will appear here"
-                        )
-                
-                build_btn.click(
-                    build_database,
-                    inputs=[db_folder, db_name, db_prompt, use_direct_pe],
-                    outputs=[db_status]
-                )
-            
-            # Tab 2: Search Similar
-            with gr.TabItem("🔎 Search Similar"):
-                gr.Markdown("## Search for similar regions in your database")
-                input_image = gr.Image(label="Upload Image", type="pil")
-                text_prompt = gr.Textbox(label="Detection Prompt", value="person . car . building")
-                use_direct_pe = gr.Checkbox(label="Use Direct PE (no region detection)", value=False)
-                detect_button = gr.Button("Detect Regions / Extract Embeddings")
-                detected_viz = gr.Image(label="Detected Regions", type="pil")
-                region_selector = gr.Dropdown(label="Select Region to Search", choices=[], visible=False)
-                detect_status = gr.Markdown()
-                
-                # After detection, update region_selector and detected_viz
-                def detect_and_extract_with_state(image, text_prompt, use_direct_pe):
-                    viz, status, region_options, metadata = detect_and_extract(image, text_prompt, use_direct_pe)
-                    # region_options is a gr.update object
-                    return viz, status, region_options, image
-                
-                detect_button.click(
-                    detect_and_extract_with_state,
-                    inputs=[input_image, text_prompt, use_direct_pe],
-                    outputs=[detected_viz, detect_status, region_selector, detected_image_state]
-                )
-                
-                # Update visualization when region is selected
-                region_selector.change(
-                    update_region_visualization,
-                    inputs=[region_selector, detected_image_state],
-                    outputs=detected_viz
-                )
-                
-                # Search controls
-                similarity_threshold = gr.Slider(minimum=0.0, maximum=1.0, value=0.7, step=0.05, label="🎚️ Similarity Threshold")
-                max_results = gr.Dropdown(choices=[3, 5, 10, 20], value=5, label="📊 Max Results")
-                search_btn = gr.Button("🎯 Search Database", variant="secondary")
-                search_results = gr.Textbox(label="🔍 Search Results", lines=10)
-                similar_image_selector = gr.Dropdown(label="Select Similar Image", choices=[], value=None, info="Choose an image to view")
-                similar_image_display = gr.Image(label="Selected Similar Image", type="pil")
-                thumbnail_gallery = gr.Gallery(label="Search Results Thumbnails", columns=3, height=300)
-                
-                # Search uses selected region
-                search_btn.click(
-                    search_database,
-                    inputs=[similarity_threshold, max_results, similar_images_state, region_selector],
-                    outputs=[search_results, similar_image_selector, similar_image_display, similar_images_state, thumbnail_gallery]
-                )
-                
-                similar_image_selector.change(
-                    update_similar_image,
-                    inputs=[similar_image_selector, similar_images_state],
-                    outputs=[similar_image_display]
-                )
-            
-            # Tab 3: Database Management
-            with gr.TabItem("⚙️ Database Management"):
-                gr.Markdown("## Manage your databases")
-                
-                with gr.Row():
-                    with gr.Column():
-                        # Automatically list available databases in a dropdown
-                        available_databases = reverso.list_databases()
-                        db_selector = gr.Dropdown(
-                            label="Select Database",
-                            choices=available_databases,
-                            value=None,
-                            info="Choose a database to manage"
-                        )
-                        
-                        with gr.Row():
-                            load_btn = gr.Button("📂 Load Database", variant="primary")
-                            delete_btn = gr.Button("🗑️ Delete Database", variant="stop")
-                            unlock_btn = gr.Button("🔓 Unlock Database", variant="secondary")
-                            reload_btn = gr.Button("🔄 Reload List", variant="secondary")
-                        
-                        db_status = gr.Textbox(
-                            label="Status",
-                            lines=3,
-                            info="Operation status will appear here"
-                        )
-                
-                # Connect functions
-                load_btn.click(
-                    load_selected_database,
-                    inputs=[db_selector],
-                    outputs=[db_status]
-                )
-                
-                delete_btn.click(
-                    delete_selected_database,
-                    inputs=[db_selector],
-                    outputs=[db_status]
-                )
-                
-                unlock_btn.click(
-                    unlock_selected_database,
-                    inputs=[db_selector],
-                    outputs=[db_status]
-                )
-
-                def reload_database_list():
-                    """Reload the list of available databases"""
-                    databases = reverso.list_databases()
-                    if not databases:
-                        return gr.update(choices=[], value=None), "No databases found"
-                    return gr.update(choices=databases, value=databases[0]), f"✅ Reloaded database list. Found {len(databases)} databases."
-                
-                reload_btn.click(
-                    reload_database_list,
-                    inputs=[],
-                    outputs=[db_selector, db_status]
-                )
-            
-            # Tab 4: About
-            with gr.TabItem("ℹ️ About"):
-                gr.Markdown("""
-                ## 🎯 Simple Revers-o: Visual Investigation Tool
-                
-                **Powered by cutting-edge AI research:**
-                - 🤖 **GroundedSAM**: Zero-shot object detection with natural language
-                - 🧠 **PE-Core-L14-336**: Meta's Perception Encoder (optimized for investigation)
-                - 🎬 **PySceneDetect**: Intelligent scene detection for video keyframe extraction
-                
-                
-                ### 📋 How to Use:
-                1. **Extract Video Frames**: Download videos and extract keyframes from detected scenes
-                2. **Create Database**: Point to your image folder, set detection prompts
-                3. **Search Similar**: Upload query image, detect regions, search database
-                4. **Investigate**: Find visually similar regions across your image collection
-                
-                ### 🎯 Best For:
-                - Person identification across image sets
-                - Vehicle tracking and matching  
-                - Object similarity search
-                - Scene and location matching
-                - Visual investigation workflows
-                
-                ### ⚡ Performance:
-                - **Model**: PE-Core-L14-336 (1B parameters)
-                - **Speed**: ~300ms per image (3x faster than complex approaches)
-                - **Memory**: ~4GB VRAM recommended
-                - **Quality**: Research-validated optimal layer extraction
-                
-                *Based on Meta's "Perception Encoder: The best visual embeddings are not at the output of the network" (2025)*
-                """)
-    
-    return demo
-
-# =============================================================================
-# MAIN EXECUTION
-# =============================================================================
+"""Launcher for the modular Revers-o application"""
+from reverso import create_simple_interface
 
 if __name__ == "__main__":
     print("🚀 Starting Simple Revers-o...")
-    
-    # Create and launch interface
     demo = create_simple_interface()
-    
     print("🌐 Launching interface...")
     demo.launch(
         share=False,
         server_name="127.0.0.1",
         server_port=7860,
-        show_error=True
+        show_error=True,
     )
